Add files via upload
Added everything into this zip folder as we had no other link to hand in on. and this is the only way that made sense as having to upload then hand in the word document with the link means we had to hand in everything else somewhere else but we were never informed where that is. I have emailed my lecturer however with my files as well.